third_party/blink/web_tests/platform/mac-mac13-arm64/fast/canvas/canvas-createImageBitmap-webgl-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot read properties of null (reading 'clearColor')
Check the imageBitmap of webgl.
FAIL Unexpected error: Uncaught TypeError: Cannot read properties of null (reading 'clearColor')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
